Sunteck Vasind Thane
Sunteck Vasind is a newly launched residential project that is developed by Sunteck Realty. It is located at the most promising and impeccable location in Vasind, Thane. It is situated near the Bhatsa River. Growing with unmatchable infrastructural development in a friendly neighbourhood, this is the best residential project in Thane. Sunteck Realty is a pioneer in developing landmark projects that promise residents new opportunities. In the ever-growing realty industry, they have marked significant footprints and have developed almost many projects to date. Now they are coming up Vasind, Thane to mark new heights and successful accomplishments.
This project will offer 2 BHK, 3 BHK and 4 BHK residential apartments for sale and possession will be given soon. It is the ultimate destination to live and so on because it hosts superb amenities. Sunteck Vasind Apartments In Thane is a premium housing to invest money which is located admist majestic luscious green hills. The location is extraordinary that will have various facilities surrounded by greenery, a multi-purpose court, an indoor sports hall, a cricket, a swimming pool, a gym, a badminton court, a fitness centre; a temperature-controlled indoor pool, a jogging track, etc. These floor plans are the best and made with keeping optimized space usage in mind. All the apartments are absolute pleasures to live in and especially all corners of the unit have access to light and are ventilated.
Now, you can gather and enjoy family time in spacious living spaces, feel the comfort and soothe up your senses in the exclusive bedrooms or watch views. The impressive designs, 24*7 safety and security, open spaces and splendid views of the luscious greenery are something that makes it a top residential project in Vasind, Thane. 
To know more - https://bit.ly/3Ycdqu3